Reorganize imports and publications
This commit...

1. publishes `unittest.TestCase` as `unittesting.TestCase` for convenience.
2. imports `expectedFailure` via `core.py38.case` for API compatibility.
