+++ unittesting/core/loader.py
@@ -1,5 +1,6 @@
 from unittest import TestSuite
 from unittest import TestLoader
+
 from . import DeferrableTestSuite
