--- unittesting/core/py38/__init__.py
@@ -1,3 +1,0 @@
-from . import case  # noqa: F401
-from . import runner  # noqa: F401
-from . import suite  # noqa: F401
