+++ unittesting/helpers/override_preferences_test_cast.py
@@ -3,7 +3,7 @@ import json
 import shutil
 import sublime
 
-from .. import DeferrableTestCase
+from ..core import DeferrableTestCase
 
 
 class OverridePreferencesTestCase(DeferrableTestCase):
