+++ unittesting/helpers/temp_directory_test_case.py
@@ -3,7 +3,7 @@ import shutil
 import sublime
 import tempfile
 
-from .. import DeferrableTestCase
+from ..core import DeferrableTestCase
 
 
 class TempDirectoryTestCase(DeferrableTestCase):
